% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{add_pmline}
\alias{add_pmline}
\title{Add PMTile outline line layers to a map}
\usage{
add_pmline(m, outlines)
}
\arguments{
\item{m}{map or map_proxy}

\item{outlines}{list of specs, each with: url, source_layer, and optionally
id, source_id, line_color, line_width, line_opacity, filter, before_id}
}
\value{
map (pipeable)
}
\description{
Add PMTile outline line layers to a map
}
\concept{viz}
