% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_bundle.R
\name{app_datasets}
\alias{app_datasets}
\title{The release's datasets, normalised}
\usage{
app_datasets(con)
}
\arguments{
\item{con}{a DBI connection}
}
\value{
a list of dataset objects
}
\description{
The release's datasets, normalised
}
\concept{app}
