% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_bundle.R
\name{app_flower_default}
\alias{app_flower_default}
\title{Versioned flower-plot defaults, per subregion}
\usage{
app_flower_default(con)
}
\arguments{
\item{con}{a DBI connection}
}
\value{
a named list \verb{subregion_key -> list of \{component, score\}}
}
\description{
Versioned flower-plot defaults, per subregion
}
\concept{app}
