% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_bundle.R
\name{app_schema_path}
\alias{app_schema_path}
\alias{app_schema_names}
\title{Path to a published \verb{app/} JSON Schema}
\usage{
app_schema_path(name)

app_schema_names()
}
\arguments{
\item{name}{one of \code{boot}, \code{taxa}, \code{taxon}, \code{alias}, \code{table}, \code{manifest}}
}
\value{
a file path

for \code{app_schema_names()}, the available schema names
}
\description{
Path to a published \verb{app/} JSON Schema
}
\concept{app}
