% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{atlas_is_preview}
\alias{atlas_is_preview}
\title{Is this process the preview (signed-in) app instance?}
\usage{
atlas_is_preview()
}
\value{
logical scalar: \code{MS_PREVIEW} is \code{1}/\code{true}
}
\description{
Is this process the preview (signed-in) app instance?
}
\concept{version}
