% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_store.R
\name{content_url}
\alias{content_url}
\title{Public URL of a stored COG}
\usage{
content_url(grid_id, content_hash, base = atlas_base_url(), ext = "tif")
}
\arguments{
\item{grid_id, content_hash}{see \code{\link[=content_key]{content_key()}}}

\item{base}{atlas base URL from \code{\link[=atlas_base_url]{atlas_base_url()}}}

\item{ext}{file extension}
}
\value{
absolute \verb{https://} URL(s)
}
\description{
Public URL of a stored COG
}
\concept{cog_store}
