% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{HEX_RES}
\alias{HEX_RES}
\title{H3 resolution of the v8 sampling grid (~5.16 km^2 per hexagon)}
\usage{
HEX_RES
}
\description{
H3 resolution of the v8 sampling grid (~5.16 km^2 per hexagon)
}
\concept{dormant}
\concept{hex}
