% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_key.R
\name{mdl_key_parse}
\alias{mdl_key_parse}
\title{Parse \code{mdl_key}(s) into components}
\usage{
mdl_key_parse(mdl_key)
}
\arguments{
\item{mdl_key}{character \code{mdl_key}(s)}
}
\value{
a tibble with columns \code{mdl_key}, \code{dataset_key}, \code{sp_id}, \code{interval},
\code{taxon_authority}, \code{taxon_id} (\code{NA} where not applicable)
}
\description{
Inverse of \code{\link[=mdl_key_raw]{mdl_key_raw()}} / \code{\link[=mdl_key_merged]{mdl_key_merged()}}: splits on \code{|} into
\code{dataset_key}, \code{sp_id}, \code{interval}; for \code{ms_merge} keys the \code{sp_id}
(\code{AUTHORITY:id}) is further split into \code{taxon_authority} + \code{taxon_id}.
}
\examples{
mdl_key_parse(c("am|Fis-29291", "gm|1234|01", "ms_merge|WORMS:137209"))
}
\concept{mdl_key}
