% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{mean_score}
\alias{mean_score}
\title{Weighted mean of component scores}
\usage{
mean_score(d_scores)
}
\arguments{
\item{d_scores}{tibble from \code{\link[=scores_for_cells]{scores_for_cells()}}}
}
\value{
a numeric scalar
}
\description{
Convenience wrapper returning the weighted mean of the \code{score}
column from \code{\link[=scores_for_cells]{scores_for_cells()}}, weighted by \code{even}.
}
\concept{calc}
