% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_asset.R
\name{native_hash_file}
\alias{native_hash_file}
\title{MD5 of an object's bytes (the store hash for an already-built object)}
\usage{
native_hash_file(path)
}
\arguments{
\item{path}{local file(s)}
}
\value{
32-hex MD5(s)
}
\description{
MD5 of an object's bytes (the store hash for an already-built object)
}
\concept{cog_store}
