% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_store.R
\name{native_vector_hashes}
\alias{native_vector_hashes}
\title{Content hash of every model in a multi-model sf}
\usage{
native_vector_hashes(x, ...)
}
\arguments{
\item{x}{an sf with a \code{mdl_key} column}

\item{...}{passed to \code{\link[=native_vector_hash]{native_vector_hash()}}}
}
\value{
data frame \code{mdl_key}, \code{n_features}, \code{content_hash}
}
\description{
Content hash of every model in a multi-model sf
}
\concept{cog_store}
