% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{spp_db_con}
\alias{spp_db_con}
\title{Connect to species taxonomy DuckDB}
\usage{
spp_db_con(read_only = TRUE)
}
\arguments{
\item{read_only}{logical; open in read-only mode (default: TRUE)}
}
\value{
DBI connection object
}
\description{
Open a DBI connection to the species taxonomy reference DuckDB database.
}
\concept{db}
