% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{stac_dataset_collection}
\alias{stac_dataset_collection}
\title{Dataset Collection node}
\usage{
stac_dataset_collection(ds, cfg, item_files = character())
}
\arguments{
\item{ds}{one-row data.frame from the \code{dataset} table}

\item{cfg}{config from \code{\link[=stac_cfg]{stac_cfg()}}}
}
\value{
STAC Collection as a list
}
\description{
Dataset Collection node
}
\concept{stac}
