% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{stac_root_catalog}
\alias{stac_root_catalog}
\title{Root Catalog node}
\usage{
stac_root_catalog(cfg)
}
\arguments{
\item{cfg}{config from \code{\link[=stac_cfg]{stac_cfg()}}}
}
\value{
STAC Catalog as a list
}
\description{
Root Catalog node
}
\concept{stac}
