% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{stac_version_collection}
\alias{stac_version_collection}
\title{Version Collection node (parent of the dataset Collections)}
\usage{
stac_version_collection(ds_keys, cfg)
}
\arguments{
\item{ds_keys}{character vector of dataset keys}

\item{cfg}{config from \code{\link[=stac_cfg]{stac_cfg()}}}
}
\value{
STAC Collection as a list
}
\description{
Version Collection node (parent of the dataset Collections)
}
\concept{stac}
