% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_area.R
\name{view_zoom}
\alias{view_zoom}
\title{Zoom that frames a given angular radius}
\usage{
view_zoom(radius_deg)
}
\arguments{
\item{radius_deg}{angular radius from the centre, degrees}
}
\value{
a MapLibre zoom, clamped to \verb{[1.2, 5]}
}
\description{
Zoom that frames a given angular radius
}
\concept{study_area}
