% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_set.R
\name{zone_set_key}
\alias{zone_set_key}
\title{Compose (and validate) a zone-set key}
\usage{
zone_set_key(zone_type, vintage)
}
\arguments{
\item{zone_type}{one of \code{planarea}, \code{programarea}, \code{ecoregion}, \code{subregion}}

\item{vintage}{\code{YYYY-MM}}
}
\value{
\code{"{zone_type}_{YYYY-MM}"}
}
\description{
Compose (and validate) a zone-set key
}
\concept{zone_set}
