% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{cells_in_polygon}
\alias{cells_in_polygon}
\title{Cells intersecting a polygon}
\usage{
cells_in_polygon(poly, r_cell_id)
}
\arguments{
\item{poly}{an sf polygon (assumed or transformable to EPSG:4326)}

\item{r_cell_id}{a single-layer \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} of integer cell ids}
}
\value{
a tibble with columns \code{cell_id} (integer) and \code{pct_covered} (0-100)
}
\description{
Given an sf polygon and a cell-id raster, return a tibble of
intersecting \code{cell_id} with \code{pct_covered} (0-100). The cell raster
uses 0-360 longitudes, so the input polygon is transformed and
shifted accordingly.
}
\concept{calc}
