% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_centroids}
\alias{hex_centroids}
\title{Centroids (lon/lat) for H3 string ids}
\usage{
hex_centroids(hex_id, con = NULL)
}
\arguments{
\item{hex_id}{character vector of H3 string ids}

\item{con}{optional DuckDB connection (a temporary one is used if \code{NULL})}
}
\value{
tibble(hex_id, lon, lat)
}
\description{
Centroids (lon/lat) for H3 string ids
}
\concept{hex}
