% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_centroids}
\alias{hex_centroids}
\title{Centroids (lon/lat) for BIGINT H3 ids}
\usage{
hex_centroids(hex_id, con = NULL)
}
\arguments{
\item{hex_id}{integer64/numeric vector of BIGINT H3 ids}

\item{con}{optional DuckDB connection (a temporary integer64 one if \code{NULL})}
}
\value{
tibble(hex_id, lon, lat)
}
\description{
Centroids (lon/lat) for BIGINT H3 ids
}
\concept{hex}
