% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_id_to_string}
\alias{hex_id_to_string}
\title{Hex-string form of BIGINT H3 ids (for display)}
\usage{
hex_id_to_string(hex_id, con = NULL)
}
\arguments{
\item{hex_id}{integer64/numeric vector of BIGINT H3 ids}

\item{con}{optional DuckDB connection (a temporary one is used if \code{NULL})}
}
\value{
character vector of H3 strings (e.g. "8729a411cffffff")
}
\description{
Hex-string form of BIGINT H3 ids (for display)
}
\concept{hex}
