% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_asset.R
\name{native_key}
\alias{native_key}
\title{Object key of a stored original surface}
\usage{
native_key(ds_key, hash, type = c("cog", "pmtiles"))
}
\arguments{
\item{ds_key}{dataset key (\code{"am"}, \code{"bl"}, \code{"rng_iucn"}, ...)}

\item{hash}{content hash (16 hex = source content, 32 hex = object-bytes MD5)}

\item{type}{\code{"cog"} (\code{.tif}) or \code{"pmtiles"}}
}
\value{
relative key(s), e.g. \code{"native/bl/3f2a....pmtiles"}
}
\description{
Object key of a stored original surface
}
\concept{cog_store}
