% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sdm_db_con}
\alias{sdm_db_con}
\title{Connect to SDM DuckDB}
\usage{
sdm_db_con(version = "2026", read_only = FALSE)
}
\arguments{
\item{version}{version date suffix (default: "2026")}

\item{read_only}{logical; open in read-only mode (default: FALSE)}
}
\value{
DBI connection object
}
\description{
Open a DBI connection to the species distribution model DuckDB database.
}
\concept{db}
