% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sdm_db_path}
\alias{sdm_db_path}
\title{Path to SDM DuckDB}
\usage{
sdm_db_path(version = "v3")
}
\arguments{
\item{version}{version suffix (default: "v3")}
}
\value{
character path to the DuckDB file
}
\description{
Get the file path to the species distribution model DuckDB database.
}
\concept{db}
